+++ src/gadgets/tests/helpers.rs
@@ -4,27 +4,28 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use crate::{double_key, gadgets, single_key};
-
+use crate::gadgets;
+use crate::{Proof, PublicKeyPair, Signature};
 use anyhow::Result;
+use dusk_pki::{PublicKey, SecretKey};
 use dusk_plonk::constraint_system::ecc::Point;
 use dusk_plonk::prelude::*;
 use rand::rngs::StdRng;
 
 #[derive(Debug, Clone)]
 pub struct SingleKeyCircuit {
-    signature: single_key::Signature,
-    sk: single_key::SecretKey,
-    pk: single_key::PublicKey,
+    signature: Signature,
+    sk: SecretKey,
+    pk: PublicKey,
     message: BlsScalar,
     pub_inputs: Vec<PublicInput>,
 }
 
 impl SingleKeyCircuit {
     pub fn new(
-        signature: single_key::Signature,
-        sk: single_key::SecretKey,
-        pk: single_key::PublicKey,
+        signature: Signature,
+        sk: SecretKey,
+        pk: PublicKey,
         message: BlsScalar,
     ) -> Self {
         Self {
@@ -72,24 +73,24 @@ impl Circuit<'_> for SingleKeyCircuit {
 
 #[derive(Debug, Clone)]
 pub struct DoubleKeyCircuit {
-    signature: double_key::Signature,
-    sk: double_key::SecretKey,
-    pk: double_key::PublicKeyPair,
+    proof: Proof,
+    sk: SecretKey,
+    pkp: PublicKeyPair,
     message: BlsScalar,
     pub_inputs: Vec<PublicInput>,
 }
 
 impl DoubleKeyCircuit {
     pub fn new(
-        signature: double_key::Signature,
-        sk: double_key::SecretKey,
-        pk: double_key::PublicKeyPair,
+        proof: Proof,
+        sk: SecretKey,
+        pkp: PublicKeyPair,
         message: BlsScalar,
     ) -> Self {
         Self {
-            signature,
+            proof,
             sk,
-            pk,
+            pkp,
             message,
             pub_inputs: vec![],
         }
@@ -99,15 +100,23 @@ impl DoubleKeyCircuit {
 impl Circuit<'_> for DoubleKeyCircuit {
     #[allow(non_snake_case)]
     fn gadget(&mut self, composer: &mut StandardComposer) -> Result<()> {
-        let R = Point::from_private_affine(composer, self.signature.R().into());
+        let R = Point::from_private_affine(
+            composer,
+            (self.proof.R().0).0.as_ref().into(),
+        );
         let R_prime = Point::from_private_affine(
             composer,
-            self.signature.R_prime().into(),
+            (self.proof.R().0).1.as_ref().into(),
         );
-        let u = composer.add_input(self.signature.u().clone().into());
-        let PK = Point::from_private_affine(composer, self.pk.PK().into());
-        let PK_prime =
-            Point::from_private_affine(composer, self.pk.PK_prime().into());
+        let u = composer.add_input(self.proof.u().clone().into());
+        let PK = Point::from_private_affine(
+            composer,
+            (self.pkp.0).0.as_ref().into(),
+        );
+        let PK_prime = Point::from_private_affine(
+            composer,
+            (self.pkp.0).1.as_ref().into(),
+        );
         let message = composer.add_input(self.message);
 
         gadgets::double_key_verify(
@@ -115,12 +124,12 @@ impl Circuit<'_> for DoubleKeyCircuit {
         );
 
         let s = composer.circuit_size();
-        let pk = self.pk.PK().into();
+        let pk = (self.pkp.0).0.as_ref().into();
         self.pub_inputs.push(PublicInput::AffinePoint(pk, s, s + 1));
         composer.assert_equal_public_point(PK, pk);
 
         let s = composer.circuit_size();
-        let pk_prime = self.pk.PK_prime().into();
+        let pk_prime = (self.pkp.0).1.as_ref().into();
         self.pub_inputs
             .push(PublicInput::AffinePoint(pk_prime, s, s + 1));
         composer.assert_equal_public_point(PK_prime, pk_prime);
@@ -145,32 +154,23 @@ impl Circuit<'_> for DoubleKeyCircuit {
 
 pub fn gen_single(
     rng: &mut StdRng,
-) -> (
-    single_key::SecretKey,
-    single_key::PublicKey,
-    BlsScalar,
-    single_key::Signature,
-) {
-    let sk = single_key::SecretKey::new(rng);
-    let pk = single_key::PublicKey::from(&sk);
+) -> (SecretKey, PublicKey, BlsScalar, Signature) {
+    let sk = SecretKey::random(rng);
+    let pk = PublicKey::from(&sk);
     let message = BlsScalar::random(rng);
-    let signature = sk.sign(rng, message);
+    let signature = Signature::new(&sk, rng, message);
 
     (sk, pk, message, signature)
 }
 
 pub fn gen_double(
     rng: &mut StdRng,
-) -> (
-    double_key::SecretKey,
-    double_key::PublicKeyPair,
-    BlsScalar,
-    double_key::Signature,
-) {
-    let sk = double_key::SecretKey::new(rng);
-    let pk = double_key::PublicKeyPair::from(&sk);
-    let message = BlsScalar::random(rng);
-    let signature = sk.sign(rng, message);
+) -> (SecretKey, PublicKeyPair, BlsScalar, Proof) {
+    let sk = SecretKey::random(rng);
+    let pkp = PublicKeyPair::from(sk);
 
-    (sk, pk, message, signature)
+    let message = BlsScalar::random(rng);
+    let proof = Proof::new(&sk, rng, message);
+
+    (sk, pkp, message, proof)
 }