+++ tests/single_key_tests.rs
@@ -5,45 +5,42 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use dusk_bls12_381::BlsScalar;
-use schnorr::single_key::{PublicKey, SecretKey};
+use dusk_bytes::Serializable;
+use dusk_pki::{PublicKey, SecretKey};
+use schnorr::Signature;
 
 #[test]
 // TestSignVerify
 fn sign_verify() {
-    let sk = SecretKey::new(&mut rand::thread_rng());
+    let sk = SecretKey::random(&mut rand::thread_rng());
     let message = BlsScalar::random(&mut rand::thread_rng());
     let pk = PublicKey::from(&sk);
 
-    let sig = sk.sign(&mut rand::thread_rng(), message);
-    let b = sig.verify(&pk, message);
+    let sig = Signature::new(&sk, &mut rand::thread_rng(), message);
 
-    assert!(b.is_ok());
+    assert!(sig.verify(&pk, message));
 }
 
 #[test]
 // Test to see failure with random Public Key
 fn test_wrong_keys() {
-    let sk = SecretKey::new(&mut rand::thread_rng());
-    let wrong_sk = SecretKey::new(&mut rand::thread_rng());
+    let sk = SecretKey::random(&mut rand::thread_rng());
+    let wrong_sk = SecretKey::random(&mut rand::thread_rng());
     let message = BlsScalar::random(&mut rand::thread_rng());
 
-    let sig = sk.sign(&mut rand::thread_rng(), message);
+    let sig = Signature::new(&sk, &mut rand::thread_rng(), message);
 
     // Derive random public key
     let pk = PublicKey::from(&wrong_sk);
-    let b = sig.verify(&pk, message);
 
-    assert!(b.is_err());
+    assert!(!sig.verify(&pk, message));
 }
 
 #[test]
 fn to_from_bytes() {
-    let sk = SecretKey::new(&mut rand::thread_rng());
-    assert_eq!(sk, SecretKey::from_bytes(&sk.to_bytes()).unwrap());
+    let sk = SecretKey::random(&mut rand::thread_rng());
     let message = BlsScalar::random(&mut rand::thread_rng());
-    let pk = PublicKey::from(&sk);
-    assert_eq!(pk, PublicKey::from_bytes(&pk.to_bytes()).unwrap());
-    let sig = sk.sign(&mut rand::thread_rng(), message);
-    use schnorr::single_key::Signature;
+
+    let sig = Signature::new(&sk, &mut rand::thread_rng(), message);
     assert_eq!(sig, Signature::from_bytes(&sig.to_bytes()).unwrap());
 }